# apps/api-gateway/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Servicienta API Gateway
  version: 0.1.0
  description: |
    OpenAPI specification for the current Express API Gateway.

    Notes:
    - Most application responses use the `{ data: ... }` envelope.
    - Error responses use `{ error, requestId }`.
    - `/health` is the only endpoint that does not use the standard envelope.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/orders:
    post:
      tags: [Orders]
      summary: Create order
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderInput'
            examples:
              createOrder:
                value:
                  flow_type: client_selects
                  description: Heladera no enfría
                  service_address_text: Calle Falsa 123
                  service_lat: -34.6037
                  service_lng: -58.3816
                  address_notes: Timbre 2, piso 4
      responses:
        '201':
          $ref: '#/components/responses/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                clientRoleRequired:
                  value:
                    error: Client role required
                    requestId: req_123
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/orders/current:
    get:
      tags: [Orders]
      summary: List current client orders
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated current client orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrdersResponse'
              examples:
                success:
                  value:
                    data:
                      items:
                        - id: eaf86abc-ae80-486a-a0cf-6cacbe761111
                          client_id: 1f0fe2c1-e1ff-4f4a-a965-14bc3cb99631
                          technician_id: 0f095605-b76a-45c5-8bd6-0551d3845d22
                          status: pending
                          flow_type: client_selects
                          description: Heladera no enfría
                          service_address_text: Calle Falsa 123
                          service_lat: -34.6037
                          service_lng: -58.3816
                          address_notes: Timbre 2, piso 4
                          zone_slug: caballito
                          appliance_type_slug: heladeras
                          created_at: '2026-05-03T14:00:00.000Z'
                          updated_at: '2026-05-03T14:00:00.000Z'
                      pagination:
                        page: 1
                        pageSize: 25
                        total: 1
                        totalPages: 1
                      summary:
                        totalOrders: 1
                        pendingOrders: 1
                        acceptedOrders: 0
                        cancelledOrders: 0
                        inProgressOrders: 0
                        completedTechOrders: 0
                        completedOrders: 0
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                clientRoleRequired:
                  value:
                    error: Client role required
                    requestId: req_123
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get current client order by id
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          $ref: '#/components/responses/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                clientRoleRequired:
                  value:
                    error: Client role required
                    requestId: req_123
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                orderNotFound:
                  value:
                    error: Order not found
                    requestId: req_123
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/orders/{orderId}/cancel:
    post:
      tags: [Orders]
      summary: Cancel current order
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          $ref: '#/components/responses/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/orders/{orderId}/accept:
    post:
      tags: [Orders]
      summary: Accept assigned order as technician and create pending operation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          $ref: '#/components/responses/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/orders:
    get:
      tags: [Orders]
      summary: List orders for admin
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: flowType
          in: query
          schema:
            $ref: '#/components/schemas/OrderFlowType'
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated admin orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAdminOrdersResponse'
              examples:
                success:
                  value:
                    data:
                      items:
                        - id: eaf86abc-ae80-486a-a0cf-6cacbe761111
                          client_id: 1f0fe2c1-e1ff-4f4a-a965-14bc3cb99631
                          client_email: client@example.com
                          client_name: Kiara
                          client_surname: Pérez
                          client_status: ACTIVE
                          technician_id: 0f095605-b76a-45c5-8bd6-0551d3845d22
                          status: in_progress
                          flow_type: client_selects
                          description: Heladera no enfría
                          service_address_text: Calle Falsa 123
                          service_lat: -34.6037
                          service_lng: -58.3816
                          address_notes: Timbre 2, piso 4
                          zone_slug: caballito
                          appliance_type_slug: heladeras
                          created_at: '2026-05-03T14:00:00.000Z'
                          updated_at: '2026-05-03T14:05:00.000Z'
                      pagination:
                        page: 1
                        pageSize: 25
                        total: 1
                        totalPages: 1
                      summary:
                        totalOrders: 1
                        pendingOrders: 0
                        acceptedOrders: 0
                        cancelledOrders: 0
                        inProgressOrders: 1
                        completedTechOrders: 0
                        completedOrders: 0
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order by id for admin
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          $ref: '#/components/responses/AdminOrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                orderNotFound:
                  value:
                    error: Order not found
                    requestId: req_123
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        $ref: '#/components/schemas/UsersPageSize'
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    OrderResponse:
      description: Order response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OrderResponse'
          examples:
            success:
              value:
                data:
                  id: eaf86abc-ae80-486a-a0cf-6cacbe761111
                  client_id: 1f0fe2c1-e1ff-4f4a-a965-14bc3cb99631
                  technician_id: 0f095605-b76a-45c5-8bd6-0551d3845d22
                  status: pending
                  flow_type: client_selects
                  description: Heladera no enfría
                  service_address_text: Calle Falsa 123
                  service_lat: -34.6037
                  service_lng: -58.3816
                  address_notes: Timbre 2, piso 4
                  zone_slug: caballito
                  appliance_type_slug: heladeras
                  created_at: '2026-05-03T14:00:00.000Z'
                  updated_at: '2026-05-03T14:00:00.000Z'
    AdminOrderResponse:
      description: Admin order response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminOrderResponse'
          examples:
            success:
              value:
                data:
                  id: eaf86abc-ae80-486a-a0cf-6cacbe761111
                  client_id: 1f0fe2c1-e1ff-4f4a-a965-14bc3cb99631
                  client_email: client@example.com
                  client_name: Kiara
                  client_surname: Pérez
                  client_status: ACTIVE
                  technician_id: 0f095605-b76a-45c5-8bd6-0551d3845d22
                  status: in_progress
                  flow_type: client_selects
                  description: Heladera no enfría
                  service_address_text: Calle Falsa 123
                  service_lat: -34.6037
                  service_lng: -58.3816
                  address_notes: Timbre 2, piso 4
                  zone_slug: caballito
                  appliance_type_slug: heladeras
                  created_at: '2026-05-03T14:00:00.000Z'
                  updated_at: '2026-05-03T14:05:00.000Z'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidPage:
              value:
                error: Invalid page
                requestId: req_123
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                error: Missing bearer token
                requestId: req_123
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                error: Admin role required
                requestId: req_123
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal:
              value:
                error: Internal server error
                requestId: req_123
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: string
        requestId:
          type:
            - string
            - 'null'
...
    UserStatus:
      type: string
      enum: [ACTIVE, DELETED]
    OrderStatus:
      type: string
      enum:
        [pending, accepted, cancelled, in_progress, completed_tech, completed]
    OrderFlowType:
      type: string
      enum: [client_selects, tech_applies]
...
    UsersPageSize:
      type: integer
      enum: [25, 50, 100]
...
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          $ref: '#/components/schemas/UsersPageSize'
        total:
          type: integer
        totalPages:
          type: integer
...
    Order:
      type: object
      required:
        - id
        - client_id
        - client_name
        - client_surname
        - client_phone
        - client_whatsapp_phone
        - technician_id
        - technician_name
        - technician_surname
        - technician_phone
        - technician_whatsapp_phone
        - status
        - flow_type
        - description
        - service_address_text
        - service_lat
        - service_lng
        - address_notes
        - zone_slug
        - appliance_type_slug
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        client_name:
          type:
            - string
            - 'null'
        client_surname:
          type:
            - string
            - 'null'
        client_phone:
          type:
            - string
            - 'null'
        client_whatsapp_phone:
          type:
            - string
            - 'null'
        technician_id:
          type:
            - string
            - 'null'
          format: uuid
        technician_name:
          type:
            - string
            - 'null'
        technician_surname:
          type:
            - string
            - 'null'
        technician_phone:
          type:
            - string
            - 'null'
        technician_whatsapp_phone:
          type:
            - string
            - 'null'
        status:
          $ref: '#/components/schemas/OrderStatus'
        flow_type:
          $ref: '#/components/schemas/OrderFlowType'
        description:
          type: string
        service_address_text:
          type: string
        service_lat:
          type:
            - number
            - 'null'
        service_lng:
          type:
            - number
            - 'null'
        address_notes:
          type:
            - string
            - 'null'
        zone_slug:
          type:
            - string
            - 'null'
        appliance_type_slug:
          type:
            - string
            - 'null'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AdminOrder:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          required:
            - client_email
            - client_name
            - client_surname
            - client_status
          properties:
            client_email:
              type: string
              format: email
            client_name:
              type:
                - string
                - 'null'
            client_surname:
              type:
                - string
                - 'null'
            client_status:
              $ref: '#/components/schemas/UserStatus'
    OrderResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Order'
    AdminOrderResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/AdminOrder'
    CreateOrderInput:
      type: object
      required:
        - technician_public_slug
        - zone_slug
        - appliance_type_slug
        - description
        - service_address_text
        - service_lat
        - service_lng
        - address_notes
      properties:
        technician_public_slug:
          type: string
        zone_slug:
          type: string
        appliance_type_slug:
          type: string
        description:
          type: string
        service_address_text:
          type: string
        service_lat:
          type:
            - number
            - 'null'
        service_lng:
          type:
            - number
            - 'null'
        address_notes:
          type:
            - string
            - 'null'
    PaginatedOrdersSummary:
      type: object
      required:
        [
          totalOrders,
          pendingOrders,
          acceptedOrders,
          cancelledOrders,
          inProgressOrders,
          completedTechOrders,
          completedOrders,
        ]
      properties:
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        acceptedOrders:
          type: integer
        cancelledOrders:
          type: integer
        inProgressOrders:
          type: integer
        completedTechOrders:
          type: integer
        completedOrders:
          type: integer
    PaginatedOrders:
      type: object
      required: [items, pagination, summary]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        pagination:
          $ref: '#/components/schemas/Pagination'
        summary:
          $ref: '#/components/schemas/PaginatedOrdersSummary'
    PaginatedOrdersResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/PaginatedOrders'
    PaginatedAdminOrders:
      type: object
      required: [items, pagination, summary]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminOrder'
        pagination:
          $ref: '#/components/schemas/Pagination'
        summary:
          $ref: '#/components/schemas/PaginatedOrdersSummary'
    PaginatedAdminOrdersResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/PaginatedAdminOrders'
...